pm UI: Convert `GroupPmConversationItem` to use `users`.
Remove its dependence on the old stringly-typed fields.

The existing code has relied on the selector to munge the data for UI
presentation; this now becomes the responsibility of the UI component
itself. (The logic to do so is presently inline, but is expected to be
unified with existing related logic in `recipient.js` later.)