example data: Make argument to message-constructors exact.
`Message` itself isn't exact (q.v.). This isn't much of a problem,
since we don't ever construct `Message` values... except in tests.

Tighten the typing on our `Message`-creation helpers to only accept
values that are documented to be part of `Message`.

Fix the (fortunately harmless) bug which this catches in the outbox-
reducer tests.