recipient.js: Tighten typing on `normalizeRecipients`.
... and, in so doing, fix a subtle bug.

There are fortunately only three callsites. Two of these would always
pass a list; the third, `matchRecipients`, might not.

This meant that if the name of a stream happened to be a user's
e-mail, `isMessageInNarrow` would incorrectly consider messages in
that stream to also be in the private-message narrow with that user.

Fix this by checking that a message passed to `matchRecipients` is of
type `"private"` before analyzing its `display_recipient`. Then
eliminate the relevant code.