android: Receive and handle shares from other apps.
Enables handling and receiving shares from other apps, for any
type of content (text, images, and other files).

Creates a native RN module `SharingModule` to handle the Android
part of this feature. 'ReceiveShareActivity.kt' parses the
received shared data and handles activating the JS part of the
codebase - either by sending an event to an already-running app,
or launching `MainActivity` with some initial shared data via
the native module.

Registers a new root component, `SharingRoot`. It is a dummy
component that does nothing, and `ReceiveShareActivity.kt`.
It exists because 'Sharing' is linked with launching an activity
in the Android ecosystem. But we don't always want to launch
`MainActivity` when receiving a share because it may cause two
instances of it to be opened simultaneously. We can't check
whether the app is running before an activity launches.
So, we launch this dummy component, then process the events,
identify whether the app is running or not, and handle that
as mentioned in the paragraph above, and then quickly kill this
dummy Activity. All of this happens fast enough that the
component does not even get time to render, so it's seamless.

Closes  #M117.