mailmap: Add a mailmap file.
Prompted by the migration from zulipchat.com to zulip.com.

While we're looking at this, we also take the opportunity to clean up
other names in the history.  The goal is to group all of a given
person's contributions together, and under their preferred name and
email address.

In cases where we have several names or email addresses for one person
and I don't directly know their preferences, I made guesses based on
public information, and always choosing from among the names and email
addresses they've used in the history:

 * If the contributions come from the same GitHub account, they're
   the same person.  (Alicja Raszkowska; Saket Kumar; Vishwesh Jainkuniya)

 * If one name looks like a full name and another a GitHub username,
   prefer the full name.  (Alicja Raszkowska; Anirudh Jain; Aswin G;
   Kunal Gupta; Saket Kumar; Saumya Bhatnagar; Vishwesh Jainkuniya)

 * For a preferred email address:

   * If one is at users.noreply.github.com, prefer any other.
     (Alicja Raszkowska; Saket Kumar)

   * o2.pl and go2.pl are apparently the same email provider and their
     addresses are equivalent; prefer the one found in the
     contributor's GitHub profile.  (Tomasz Kolek)

   * Prefer the one the contributor has used more often most recently.
     (Neeraj Wahi; Vishwesh Jainkuniya)