[FOR DISCUSSION] One way to get `sharedData` into ShareToStream, ShareToPm
This is just meant to show one way this could be done -- it seems
more React- and React-Navigation idiomatic. It avoids calling
`createMaterialTopTabNavigator` from within a render method or a
constructor.

This uses the alternative mentioned at the "Common Mistakes" doc
[1], with the `router` static and threading through the `navigation`
prop.

There are a few comments I would make about some of the code, not
addressed in this commit, but hopefully this will help for
conceptual clarity about the kind of thing I'm suggesting. :)

More discussion at
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60screenProps.60.20in.20react-navigation.20v2/near/888521.

[1]: https://reactnavigation.org/docs/2.x/common-mistakes/#explicitly-rendering-more-than-one-navigator