tools/test lint: Make globs passed to `prettier-eslint` absolute.
I'm not deeply familiar with the details, but there's an open issue
about an error message we're seeing in the upcoming
`prettier-eslint` update, that says "'basePath' should be an
absolute path.". That's
https://github.com/prettier/prettier-eslint-cli/issues/205.

The first reply at that thread is a workaround that works for us:
make the paths absolute. Instead of using `pwd`, as suggested there,
we reuse our existing `rootdir` variable, which looks like it's
designed for reuse for consistency.

For a discussion that will help clarify the arcane bit of Bash here:
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4120.20ESLint.206.2E8.2E0.20upgrade/near/889570.