outbox: Use `sendingDeferred` to show loading indicator for file uploads.
While we're waiting for the upload to complete, show a local-echo
outbox message with "[Uploading file...]".

Done by adding an `Outbox` object to `state.outbox` with the
placeholder text right away, but marked with `sendingDeferred` until
the upload is complete. When the upload is complete, unmark
`sendingDeferred` and put the newly available URL in the content
(via a new UPDATE_OUTBOX_MESSAGE action), and send the message.

Co-authored-by: Chris Bobbe <cbobbe@zulip.com>
Fixes: #2684