android build: Go with recommended `react-native-vector-icons` setup.
`react-native-vector-icons` gives two options for setup on Android [1]:

- Option: With Gradle (recommended)

- Option: Manually

When we first installed it in e5cbe3cc5, we...mostly went with the
latter (which is also "with Gradle": it means including the
library's Gradle project!), but we didn't follow through with an
important edit to `MainApplication.java` as part of that option, or
a (silly, as Greg points out [2]) move of some font assets into our
source tree.

In cd899bbe0, we accomplished the "With Gradle (recommended)"
option. This seems better; we've never needed to do the things
described as possible only with the "Manual" setup.

But we didn't back out the then-unnecessary code from the "Manual"
setup. So, do that now.

Greg points out [2] that they're making the Gradle setup a lot more
complicated than it really neads to be.

[1]: https://github.com/oblador/react-native-vector-icons/tree/v6.6.0#android

[2]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/RN.20autolinking/near/894795