android deps: Finish removal of `react-native-orientation`.
In 0d5a19ac4, for #4147, we forgot to remove some Android
configuration specific to `react-native-orientation` in
MainActivity.java [1], added in 3149c9d60. So, remove it.

It appears that the iOS configuration at the
`react-native-orientation` doc was never done, so there's nothing to
do for that.

Greg pointed out that this override was replicated in
ReceiveShareActivity.kt when that file was created (in 0b84717e4)
and that he was puzzled by it while reviewing #4124 and supposed it
might be generic RN boilerplate. Now that we've identified it,
remove it.

[1]: https://github.com/yamill/react-native-orientation#configuration