deps: Upgrade `react-native-webview` to ~10.0.0.
We'd like to upgrade to the latest, but versions 10.1.0 and above
don't have support for RN v0.61; minimum supported is v0.62. (Hence
the tilde.)

Breaking changes announced:

- v9: iOS: Props updates to `injectedJavaScript` are no longer
  immutable.

We don't use the `injectedJavaScript` prop, so this is fine.

- v10: gradle: The Android Gradle plugin is only required when
  opening the project stand-alone, not when it is included as a
  dependency. [...]

This doesn't sound like a description of a breaking change. As Greg
said [0], "a thing is no longer required in a case where it
previously was".

We hesitate to take an x.0.0 version -- what if there are important
bugfixes in minor/patch versions for issues introduced in that x.0.0
version? -- but they've made it a custom to post a note at the
release if they know of major regressions, telling you to use a
later version [1]. And no such note exists at v10.0.0.

As a concrete example, we were concerned that the "bug fixes"
identified in the v10.1.0 release might have been aimed at
regressions introduced in v10.0.0. But following links to the
mentioned PR and its corresponding issues (the second looks like a
resumption of the first), it looks like a fix for a bug reported all
the way back in version 5. So this isn't a reason to use v9.x.x
instead of v10.0.0.

Greg points out [2]:

"""
In a project that takes major vs. minor releases seriously, where a
minor release means cherry-picked backported changes, I'd assume
from this context that the major release introduced the issue.

(Projects like that include RN itself, and the Zulip server.)

But with this project I get the impression that their releases are
all a linear stream of snapshots from master. Also that they do the
flavor of "semantic versioning" that's basically "bump the major
version number frequently and at random". So that makes this less
clear.
"""

Ah, well.

Also, update the libdef according to changes in
`src/WebViewTypes.ts` in the `react-native-webview` repo.

[0]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/RN.2060.20upgrade.3A.20react-native-webview/near/893807
[1]: See, e.g., v10.1.0, at
     https://github.com/react-native-community/react-native-webview/releases/tag/v10.1.0.
[2]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/RN.20v0.2E61.3A.20react-native-webview/near/901944