navigation: Use `react-native-screens`.
Follow a doc [1] for optimizing memory usage and performance by
making `react-navigation` work more closely with native navigation
controls. That doc links to some performance comparisons [2], and
you can also see a talk [3] by the library's author (notable
timestamps are 18:43 and 27:04).

It means removing `react-native-screens` from a list of libraries to
exclude in autolinking.

We would also have liked to upgrade this to the latest version (or
even a non-alpha 1.0.0), but it won't be possible until we're on a
later version of React Navigation because of peer dependency
constraints in `react-native-tabs`. The react-navigation/tabs repo
has been archived and made read-only (there seems to be a different
structure in place with later `react-navigation` versions) and the
archived state of that repo has a `react-native-screens` peer
dependency of "^1.0.0 || ^1.0.0-alpha". A non-alpha 1.0.0 version
doesn't exist, and 23 is the latest. So, we'll stay there.

[1]: https://reactnavigation.org/docs/2.x/react-native-screens/
[2]: https://twitter.com/janicduplessis/status/1039979591815897088?s=21
[3]: https://www.youtube.com/watch?v=Z0Jl1KCWiag

Fixes: #4111