async [nfc]: Move `tryUntilSuccessful` to `fetchActions` and rename.
As Greg says [1], it's always been closely tied to
`doInitialFetch`'s specific needs, with its handling of 4xx errors.
We're about to make it more so, by adding timeout logic.

Also, remove two pretty trivial tests, and adapt the third (that
tests that a retry is done) to use Lolex. Add another test to
confirm that a 4xx error does indeed halt the retry loop and get
thrown.

[1]: https://github.com/zulip/zulip-mobile/issues/4165#issue-640704691