redux: Add `INITIAL_FETCH_ABORT` action.
To be dispatched when we time out attempts to connect to a server
that isn't responding, or when a server responds with a 5xx error.

It navigates to the 'accounts' screen so a user can try a different
account and server. Logging out wouldn't be good; the credentials
may be perfectly fine, and we'd like to keep them around to try
again later.

We also need to ensure `needsInitialFetch` is set to `false` [1],
while avoiding the semantics (and a navigation to the 'main' screen,
which depends on data from the initial fetch!) of
`INITIAL_FETCH_COMPLETE`. See also discussion [2].

[1]: As also noted in 7caa4d08e, `needsInitialFetch` is
     "edge-triggered":

     """
     [I]t's consulted inside a `componentDidUpdate` method, which
     means it typically will cause a `doInitialFetch` call only when
     it's *updated* to true, not if it's continuously true.
     """

     This logic seems complex and fragile, and it would be nice to
     fix that.

[2]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Stuck.20on.20loading.20screen/near/907591