url polyfill: Assign to global URL variable.
As noted at #4146, this should eliminate accidents where we use
React Native's own, buggy polyfill that they've assigned to the
global URL. If React Native didn't do that, we'd think twice before
mucking with globals -- but they do.

It also makes this polyfill easier to use because we don't have to
think about importing anything.

[1]: https://github.com/zulip/zulip-mobile/issues/4146#issuecomment-646221190