config: Add `appOwnDomains`.
These will be useful, in an upcoming commit, to grant the privilege
of the native "Sign in with Apple" flow (in which sensitive
`id_token`s will be sent to the server) only to domains that have
trust with the mobile app.

To do that securely, the base URL we'll be sending requests to
(basically what the user entered on the first screen) will be
checked to see if its host matches one of these.