home screen: Use material tabs for bottom navigation.
Use `createMaterialTopTabNavigator` for bottom navigation. The
previously used `createBottomTabNavigator` gave no feedback on
tapping the icons, which is bad UX. Infact, on low end devices,
navigating to different screens can take a noticable amount of
time, and sometimes it feels that the touch didn't register. Using
these fixes the issue.