msglist: Use absolute URL for webview assets on iOS.
In a recent commit, we exposed this from the native side of the
React Native bridge. Use it, make `baseUrl` a URL object, and remove
various hacks that are now unnecessary.

We also rename `assetsPath` to `webviewAssetsUrl`: there's no
question that it's a URL now, and it specifically refers to the
`webview/` directory in the platform-specific assets folder.

The platform-specific assets folder URL is extracted as `assetsUrl`.