realm screen: Don't set the `realm` state to the value it already has.
We just got `realm` from the state a couple of lines ago, so setting
it back to the same value here has no useful effect.  It's confusing,
so remove it.

This appears to be the legacy of an unfinished refactor a couple of
years ago in 252fed22f.  Before that, this was the place where we
invoked `fixRealm` (now `fixRealmUrl`) and set the `realm` state
property accordingly.  With that change, that responsibility moved to
within the new `SmartUrlInput`, and this line was edited into being a
no-op.