deps: Upgrade most Babel-related dependencies.
This is what we get with:

    $ yarn upgrade -S @babel

followed by the usual cleanup command:

    $ yarn yarn-deduplicate && yarn

Sadly it doesn't actually cover all our `@babel/` dependencies.  It
seems to have the exact same effect as if we pass to `yarn upgrade`
only a list of our *direct* dependencies that are in that scope:

    $ yarn upgrade -- $(tools/deps | grep @babel)

Seems like a bug in `yarn upgrade`, or at the very least a
surprising and undocumented limitation.  We'll work around it in
a followup commit.

Meanwhile, we regenerate our generated code: one of the upgraded
packages apparently causes a slight change there.