android: Handle new intents in 'onNewIntent' when activity is resumed.
The current handling for receiving 'shares' from other apps did not
work in some instances, for example, when cancelling a share in
progress and then sharing once again. This happened because, as the
Android activity-element doc[1] for the Launch Mode 'singleTask'
says:

"However, if an instance of the activity already exists, the system
routes the intent to existing instance through a call to its
onNewIntent() method, rather than creating a new one."

So, also using 'onNewIntent()' to handle the received intent,
instead of only in 'onCreate()' fixes this.

[1]: https://developer.android.com/guide/topics/manifest/activity-element

Reported-By: Greg Price <greg@zulip.com>