sharing: Involve a thunk action in navigateToSharing so it's buffered.
Currently, when sharing something to Zulip, if the app is not hydrated
(i.e. basically if it wasn't already running in the background), the
sharing screen appears and then promptly disappears.

Using a thunk action to navigate causes it to be buffered by the
'createActionBuffer' middleware until after the app's data is
hydrated, which fixes this issue.

Suggested-by: Greg Price <greg@zulip.com>