nav icons: Add optional prop "title".
While icon buttons are great for conveying purpose without text,
it can sometimes be confusing what they actually do. Most apps
allow long-pressing them to show a toast with their textual
equivalent.

This commit adds the ability to pass a 'title' prop to 'NavButton'.
On long pressing it, a toast is shown with this title string.