subscriptions reducer tests: Remove unwanted expectation from a test.
Currently, this test confirms that the reducer isn't clobbering the
details of a subscription.

It does this by making the `color` property

- present in the previous state
- missing in the subscription in the action
- present (at the same value) in the result state.

As Greg says in discussion [1], this isn't really an expectation
that we want to have:

"""
If we get an `op: 'add'` subscription event that includes a stream
we already had a subscription for, and especially if any of the data
differed, then (a) I think that means something has already gone
wrong and (b) if it does happen, we'd probably be best off doing the
opposite of what this test tested for, and clobbering our existing
clearly-stale data in favor of the new data.
"""

That being said,

"""
It's probably fine that we keep the code simple and effectively
assume this case doesn't happen, but there's no need for the test to
confirm that if it did happen we'd do the *wrong* thing.
"""

So, remove the unwanted expectation, and do it here, rather than in
the same commit where we start type-checking this file, since it's a
substantive change.

[1] https://github.com/zulip/zulip-mobile/pull/4199#discussion_r459121987