ZulipButton [nfc]: Limit the set of style properties allowed.
The use of specific style properties for callers to micromanage the
layout of the button is pretty chaotic and fragile, introducing
implicit dependencies on details of the ZulipButton implementation.

It'd be good to assimilate those usages into a more coherent set of
options provided by the ZulipButton interface explicitly.  This
would also help with making the buttons' appearance more consistent
across the app.

Until then, by listing here the properties we do use, we can at least
make it possible when working on the ZulipButton implementation to
know the scope of different ways that callers can mess with the styles.