i18n [nfc]: Add interpolation support to our `_` wrapper.
This allows us to stop using the `intl.formatMessage` API in its
most extra-verbose form.  (When there are variables to interpolate,
both `defaultMessage` and `id` are required, even though for us
they're the same thing -- because while react-intl will fall back
to `id` as the message when a translation isn't available, it will
do so literally, without any interpolation.  See discussion:
  https://github.com/zulip/zulip-mobile/pull/4175#discussion_r459261420
)